Components: Move wrapperDisplayName to elements namespace
