--- components/higher-order/wrapper-display-name/index.js
@@ -1,10 +1,0 @@
-/**
- * External dependencies
- */
-import { startCase } from 'lodash';
-
-export default function wrapperDisplayName( wrapperPrefix, WrappedComponent ) {
-	const { displayName = WrappedComponent.name || 'Component' } = WrappedComponent;
-
-	return `${ startCase( wrapperPrefix ) }(${ displayName })`;
-}
